refactor: move TS types to proper modules
