+++ tsconfig.json
@@ -23,6 +23,7 @@
   ],
   "include": [
     "next-env.d.ts",
+    "typings",
     "**/*.ts",
     "**/*.tsx"
   ]
