+++ typings/extended/react.d.ts
@@ -1,0 +1,7 @@
+import React from 'react'
+
+declare module 'react' {
+  interface DOMAttributes<T> {
+    on?: string
+  }
+}
